refactor sample app module name
