+++ create-save-map/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.esri.arcgisruntime.sample.authoringamap">
+    package="com.esri.arcgisruntime.sample.createsavemap">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -10,7 +10,7 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="com.esri.arcgisruntime.sample.authoringamap.MainActivity">
+        <activity android:name="com.esri.arcgisruntime.sample.createsavemap.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -18,7 +18,7 @@
 
             </intent-filter>
         </activity>
-        <activity android:name="com.esri.arcgisruntime.sample.authoringamap.MapSaveActivity">
+        <activity android:name="com.esri.arcgisruntime.sample.createsavemap.MapSaveActivity">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
                 <action android:name="android.intent.action.PICK" />
