+++ create-save-map/src/main/res/drawable-hdpi/ic_drawer.png
