+++ create-save-map/src/main/res/drawable-mdpi/ic_drawer.png
