+++ create-save-map/src/main/res/drawable-xxhdpi/ic_drawer.png
