+++ create-save-map/src/main/res/mipmap-hdpi/ic_launcher.png
