+++ create-save-map/src/main/res/mipmap-mdpi/ic_launcher.png
