+++ create-save-map/src/main/res/values/colors.xml
