+++ create-save-map/src/main/res/values/portal_settings.xml
