+++ settings.gradle
@@ -3,12 +3,12 @@ include ':add-graphics-renderer',
         ':arcgis-map-imagelayer-url',
         ':arcgis-tiledlayer-url',
         ':arcgis-vectortiledlayer-url',
-        ':authoring-map',
         ':change-basemaps',
         ':change-feature-layer-renderer',
         ':change-sublayer-visibility',
         ':change-viewpoint',
         ':create-geometries',
+        ':create-save-map',
         ':display-device-location',
         ':display-drawing-status',
         ':display-layer-view-state',
